chore: Update installation script

+++ install.sh
@@ -2,13 +2,15 @@
 
 function usage {
 	cat <<-EOF
-		Usage: $PROGRAM [...OPTIONS]
+		Usage: $PROGRAM [OPTIONS]
 
 		Scripts installer.
 
 		OPTIONS
 		  -b, --binpath          set the installation path for scripts, default: ~/bin
 		      --keep-extension   keep .sh extension for installed scripts
+		  -f, --force            overwrite existing files
+		  -v, --verbose          print verbose output
 		  -h, --help             print this help message
 	EOF
 }
@@ -21,7 +23,7 @@ function die {
 }
 
 function main {
-	if ! options=$(getopt -o "b:" -l "binpath:,keep-extension" -- "$@"); then
+	if ! options=$(getopt -o "hb:fv" -l "help,binpath:,keep-extension,force,verbose" -- "$@"); then
 		die
 	fi
 
@@ -29,14 +31,29 @@ function main {
 
 	while [ $# -gt 0 ]; do
 		case $1 in
+		-h | --help)
+			usage
+			exit 0
+			;;
+
 		-b | --binpath)
-			local binpath=$2
+			local OPT_BINPATH=$2
 			shift 2
 			;;
 
 		--keep-extension)
-			local keep_extension=y
-			shift 2
+			local OPT_KEEP_EXT=y
+			shift
+			;;
+
+		-v | --verbose)
+			local OPT_VERBOSE=y
+			shift
+			;;
+
+		-f | --force)
+			local OPT_FORCE=y
+			shift
 			;;
 
 		--)
@@ -54,19 +71,26 @@ function main {
 		esac
 	done
 
-	binpath="${binpath:-${HOME}/bin/}"
+	binpath="${OPT_BINPATH:-${HOME}/bin/}"
 
 	find "$(dirname "$0")/src" -type f -name '*.sh' -print0 | while read -d $'\0' filename; do
 		local out_filename="${filename##*/}"
 
-		if ! [[ $keep_extension == 'y' ]]; then
+		if ! [[ "$OPT_KEEP_EXT" == 'y' ]]; then
 			out_filename="${out_filename%.sh}"
 		fi
 
-		# Respect `~` expansion
-		eval fullpath=$(echo "$binpath/$out_filename")
+		fullpath="$(realpath "$binpath/$out_filename")"
 
-		install "$filename" "$fullpath" -m 755
+		if [[ -f "$fullpath" ]] && [[ -z "$OPT_FORCE" ]]; then
+			echo "'$fullpath' already exists but --force was not specified."
+		else
+			if [[ -n "$OPT_VERBOSE" ]]; then
+				echo "install \"$filename\" \"$fullpath\" -m 755"
+			fi
+
+			install "$filename" "$fullpath" -m 755
+		fi
 	done
 }
